// server/src/authentication/interfaces/auth.interfaces.ts
export interface RegisterDto {
  name: string;
  email: string;
  password: string;
  confirmPassword: string;
  userAgent?: string;
}

export interface LoginDto {
  email: string;
  password: string;
  userAgent?: string;
}

export interface resetPasswordDto {
  password: string;
  verificationCode: string;
}

export interface JwtPayload {
  userId: string;
  sessionId: string;
}